Refactor out date creation
