+++ lib/index.js
@@ -1,25 +1,26 @@
 var Food = require('./foods');
 var Exercise = require('./exercise');
 var Table = require('./table')
+var Dates = require('./dates')
 
 var count = 0;
 
 // generate variable names
-function inWords (num) {
-    var a = ['','one ','two ','three ','four ', 'five ','six ','seven ','eight ','nine ','ten ','eleven ','twelve ','thirteen ','fourteen ','fifteen ','sixteen ','seventeen ','eighteen ','nineteen '];
-    var b = ['', '', 'twenty','thirty','forty','fifty', 'sixty','seventy','eighty','ninety'];
-
-    if ((num = num.toString()).length > 9) return 'overflow';
-    n = ('000000000' + num).substr(-9).match(/^(\d{2})(\d{2})(\d{2})(\d{1})(\d{2})$/);
-    if (!n) return; var str = '';
-    str += (n[1] != 0) ? (a[Number(n[1])] || b[n[1][0]] + ' ' + a[n[1][1]]) + 'crore' : '';
-    str += (n[2] != 0) ? (a[Number(n[2])] || b[n[2][0]] + ' ' + a[n[2][1]]) + 'lakh' : '';
-    str += (n[3] != 0) ? (a[Number(n[3])] || b[n[3][0]] + ' ' + a[n[3][1]]) + 'thousand' : '';
-    str += (n[4] != 0) ? (a[Number(n[4])] || b[n[4][0]] + ' ' + a[n[4][1]]) + 'hundred' : '';
-    str += (n[5] != 0) ? ((str != '') ? '' : '') + (a[Number(n[5])] || b[n[5][0]] + ' ' + a[n[5][1]]) + '' : '';
-    str = str.replace(/\s+/g, '');
-    return str;
-}
+// function inWords (num) {
+//     var a = ['','one ','two ','three ','four ', 'five ','six ','seven ','eight ','nine ','ten ','eleven ','twelve ','thirteen ','fourteen ','fifteen ','sixteen ','seventeen ','eighteen ','nineteen '];
+//     var b = ['', '', 'twenty','thirty','forty','fifty', 'sixty','seventy','eighty','ninety'];
+//
+//     if ((num = num.toString()).length > 9) return 'overflow';
+//     n = ('000000000' + num).substr(-9).match(/^(\d{2})(\d{2})(\d{2})(\d{1})(\d{2})$/);
+//     if (!n) return; var str = '';
+//     str += (n[1] != 0) ? (a[Number(n[1])] || b[n[1][0]] + ' ' + a[n[1][1]]) + 'crore' : '';
+//     str += (n[2] != 0) ? (a[Number(n[2])] || b[n[2][0]] + ' ' + a[n[2][1]]) + 'lakh' : '';
+//     str += (n[3] != 0) ? (a[Number(n[3])] || b[n[3][0]] + ' ' + a[n[3][1]]) + 'thousand' : '';
+//     str += (n[4] != 0) ? (a[Number(n[4])] || b[n[4][0]] + ' ' + a[n[4][1]]) + 'hundred' : '';
+//     str += (n[5] != 0) ? ((str != '') ? '' : '') + (a[Number(n[5])] || b[n[5][0]] + ' ' + a[n[5][1]]) + '' : '';
+//     str = str.replace(/\s+/g, '');
+//     return str;
+// }
 
 function displayFoodData() {
   var foods = JSON.parse(localStorage.getItem('foods'));
@@ -34,6 +35,7 @@ function displayFoodData() {
 }
 
 function formatDateKey(){
+  var date = new Dates();
   var pageDate = document.getElementById('diary-today').innerText;
   pageDate = pageDate.replace(" ", "")
   pageDate = pageDate.replace(", ", "")
@@ -41,7 +43,7 @@ function formatDateKey(){
   var numberPattern = /\d+/g;
   dateNumbers = pageDate.match( numberPattern )[0]
 
-  pageDate = pageDate.replace(dateNumbers, inWords(dateNumbers));
+  pageDate = pageDate.replace(dateNumbers, date.inWords(dateNumbers));
   return pageDate;
 }
 
@@ -289,7 +291,6 @@ function calculateTotalCalories(){
   var exercisesTable = document.getElementById('exercise-table');
   var totalCalorieCell = document.getElementById('exercise-total-cals');
   var totalCalories = 0;
-  debugger;
 
   $(".exercise-calorie-cell").each(function() {;
       totalCalories += parseFloat(this.innerText);
